+++ src/lib.rs
@@ -80,6 +80,7 @@ pub use crate::algorithms::paths::radii::{
     eccentricity, eccentricity_weighted, eccentricity_weighted_with_mode, eccentricity_with_mode,
     radius, radius_weighted, radius_weighted_with_mode, radius_with_mode,
 };
+pub use crate::algorithms::paths::random_walk::random_walk;
 pub use crate::algorithms::properties::assortativity::{
     assortativity_degree, assortativity_degree_directed,
 };